requirements.txt
matplotlib
numpy
Pillow
pandas
scikit-learn
tensorflow
scikit-image
